backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0
supabase==2.2.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx>=0.24.0,<0.25.0
reportlab==4.0.7
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
cryptography==41.0.7
email-validator==2.1.0
strawberry-graphql[fastapi]==0.206.0
slowapi==0.1.8
redis==5.0.1
aiofiles==23.2.1
docker==7.0.0
psutil==5.9.6